refactor: use hexagonal architecture
