+++ src/main/java/com/hlag/tools/commvis/service/IScannerService.java
@@ -1,4 +1,4 @@
-package com.hlag.tools.commvis;
+package com.hlag.tools.commvis.service;
 
 public interface IScannerService {
     void scanClasspath(String rootPackageName);
